Move tests to `tests/` directory
